Only profile `batched` atm
